refactor: 쿠폰 모달을 새로운 컴포넌트로 분리
