+++ src/pages/CouponModal/CouponModal.styles.ts
@@ -1,0 +1,21 @@
+import { css } from "@emotion/react";
+
+export const RedeemInfo = css`
+  height: 16px;
+
+  display: flex;
+  gap: 4px;
+
+  font-size: var(--font-size-small);
+  font-weight: var(--font-weight-normal);
+  line-height: 16px;
+  text-align: center;
+`;
+
+export const ModalContentContainer = css`
+  width: 320px;
+
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+`;
