+++ src/pages/OrderSummary/OrderSummary.styles.ts
@@ -43,23 +43,3 @@ export const RedeemCouponButton = css`
   font-weight: var(--font-weight-bold);
   color: var(--color-dark-gray);
 `;
-
-export const RedeemInfo = css`
-  height: 16px;
-
-  display: flex;
-  gap: 4px;
-
-  font-size: var(--font-size-small);
-  font-weight: var(--font-weight-normal);
-  line-height: 16px;
-  text-align: center;
-`;
-
-export const ModalContentContainer = css`
-  width: 320px;
-
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-`;
